impemented comments #4

+++ sw/airborne/modules/optical_flow/mateksys_3901_l0x.h
@@ -81,7 +81,7 @@ struct Mateksys3901l0X {
   int32_t  motionY;
   uint8_t  distancemm_quality;
   int32_t  distancemm_temp;
-  int32_t  distancemm;
+  float  distancemm;
   float    distance_compensated;
   float    velocityX;
   float    velocityY;
